+++ frontend/components_react/guild/guild.jsx
@@ -110,7 +110,9 @@ export default class Guild extends React.Component {
     return this.state.characters.map((character) => {
       return (
         <tr className={character.character_class_name.en} key={character.id}>
-          <td className='character_link' onClick={this._goToCharacter.bind(this, character.slug)}>{character.name}</td>
+          <td className='character_link'>
+            <a href={this._defineCharacterLink(character.slug)}>{character.name}</a>
+          </td>
           <td>
             <div className="role_icons">
               <div className={`role_icon ${character.main_role_name.en}`}></div>
@@ -125,8 +127,8 @@ export default class Guild extends React.Component {
     })
   }
 
-  _goToCharacter(characterSlug) {
-    window.location.href = `${this.props.locale === 'en' ? '' : '/' + this.props.locale}/characters/${characterSlug}`
+  _defineCharacterLink(characterSlug) {
+    return `${this.props.locale === 'en' ? '' : ('/' + this.props.locale)}/characters/${characterSlug}`
   }
 
   _renderManageButtons(character) {